+++ eruption-cmd/src/main.rs
@@ -143,6 +143,7 @@ Copyright (c) 2019-2022, The Eruption Development Team
     );
 }
 
+/*
 #[cfg(debug_assertions)]
 mod thread_util {
     use crate::Result;
@@ -175,6 +176,7 @@ mod thread_util {
         Ok(())
     }
 }
+*/
 
 pub async fn async_main() -> std::result::Result<(), eyre::Error> {
     cfg_if::cfg_if! {
@@ -199,9 +201,9 @@ pub async fn async_main() -> std::result::Result<(), eyre::Error> {
     logger::initialize_logging(&env::var("RUST_LOG").unwrap_or_else(|_| "info".to_string()))?;
 
     // start the thread deadlock detector
-    #[cfg(debug_assertions)]
-    thread_util::deadlock_detector()
-        .unwrap_or_else(|e| error!("Could not spawn deadlock detector thread: {}", e));
+    // #[cfg(debug_assertions)]
+    // thread_util::deadlock_detector()
+    //     .unwrap_or_else(|e| error!("Could not spawn deadlock detector thread: {}", e));
 
     // register ctrl-c handler
     let (ctrl_c_tx, _ctrl_c_rx) = unbounded();