+++ eruption/src/plugins/mod.rs
@@ -19,6 +19,7 @@
 
 pub mod animal;
 pub mod audio;
+pub mod hwaccel;
 pub mod canvas;
 pub mod introspection;
 pub mod keyboard;
@@ -34,6 +35,7 @@ pub mod uleds;
 
 pub use animal::AnimalPlugin;
 pub use audio::AudioPlugin;
+pub use hwaccel::HwAccelerationPlugin;
 pub use canvas::CanvasPlugin;
 pub use introspection::IntrospectionPlugin;
 pub use keyboard::KeyboardPlugin;
@@ -115,6 +117,10 @@ pub fn register_plugins() -> Result<()> {
 
     let _ = plugin_manager
         .register_plugin(Box::new(AudioPlugin::new()))
+        .map_err(|_e| error!("An error occurred during initialization of the plugin"));
+
+    let _ = plugin_manager
+        .register_plugin(Box::new(HwAccelerationPlugin::new()))
         .map_err(|_e| error!("An error occurred during initialization of the plugin"));
 
     // Additional plugins